% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productionHardening.R
\name{chemistryComparisonDictionary}
\alias{chemistryComparisonDictionary}
\title{Chemistry comparison dictionary}
\usage{
chemistryComparisonDictionary()
}
\value{
Data frame defining default comparable chemistry scopes.
}
\description{
Chemistry comparison dictionary
}
