% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactoThese.R
\name{exactoThese}
\alias{exactoThese}
\title{exactoThese}
\usage{
exactoThese(
  categoratedInput,
  subsetBy = "Database",
  subsetArgs = "All",
  subsetArgs2 = NA,
  subset_input = NA
)
}
\arguments{
\item{categoratedInput}{the lists output from `categorate()`}

\item{subsetBy}{specifies the list to subset by (Database, FMCS, or Library)}

\item{subsetArgs}{additional arguments to specify which values to subset by}

\item{subsetArgs2}{additional arguments to specify which values to subset by}

\item{subset_input}{used when subsetting by FMCS information (e.g. molecular weight)}
}
\value{
A vector of chemical names that meet each user-specification. Used as input
for `mzExacto()`.
}
\description{
Takes categorated output as input and makes chemical subsets by the information in it.
}
\details{
Provides a set of search chemicals for `mzExacto()`. User gets to select
chemicals based on information generated from `categorate()`. Database
subsetting accepts both historical `Databases` output and current split
database tables from `categorate()`.
}
\examples{
exactoThese(standard_categorated, subsetBy = "Database", subsetArgs = "All")
exactoThese(standard_categorated, subsetBy = "Database", subsetArgs = "reactives")
exactoThese(standard_categorated, subsetBy = "Database", subsetArgs = "LOTUS")
exactoThese(standard_categorated, subsetBy = "Database", subsetArgs = "KEGG")
exactoThese(standard_categorated, subsetBy = "Database", subsetArgs = "FEMA")
exactoThese(standard_categorated, subsetBy = "Database", subsetArgs = "FDA_SPL")
exactoThese(standard_categorated, subsetBy = "Database", subsetArgs = c("reactives", "FEMA"))
exactoThese(standard_categorated, subsetBy = "FMCS", subsetArgs = "MW",
subsetArgs2 = "Between", subset_input = c(125, 200))
}
