% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalStandard_data.R
\docType{data}
\name{ExternalStandard_data}
\alias{ExternalStandard_data}
\title{Sample Data Set for External Standards}
\format{
## `ExternalStandard_data`
A data frame with 28 rows and 4 columns:
\describe{
 \item{Quantity}{Measured Amount of Analyte in Sample}
 \item{Component.Area}{Area/Quantity of Analyte}
 \item{Component.RT}{Retention Time of Analyte}
 \item{Replicate}{Experimental Replicate}
 }
}
\usage{
ExternalStandard_data
}
\description{
A mock data set to show the correct format for
the `ES_calibration` input when using `standardifyIt()`
with an External standard
#'
}
\keyword{datasets}
