% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_data.R
\docType{data}
\name{library_data}
\alias{library_data}
\title{Sample Chemical Library}
\format{
A data frame with 7 rows and 5 columns:
\describe{
 \item{Type.A}{The first set of chemicals}
 \item{Type.B}{The second set of chemicals}
 \item{Type.C}{The third set of chemicals}
 \item{Type.D}{The fourth set of chemicals}
 \item{Type.E}{The fifth set of chemicals}
 }
}
\usage{
library_data
}
\description{
A library of chemicals in wide format, intended to highlight
important features of the package in the vignette
}
\keyword{datasets}
