% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_data_long.R
\docType{data}
\name{library_data_long}
\alias{library_data_long}
\title{Sample Chemical Library}
\format{
A data frame with 14 rows and 2 columns:
\describe{
 \item{type}{The types of chemical}
 \item{compound}{The names of chemicals}
 }
}
\usage{
library_data_long
}
\description{
A library of chemicals in long format, intended to test `personalLib()`
}
\keyword{datasets}
