% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzExacto.R
\name{mzExacto}
\alias{mzExacto}
\title{mzExacto}
\usage{
mzExacto(data_in, chemicals, decontaminate = T)
}
\arguments{
\item{data_in}{the list output from `spreadOut()`}

\item{chemicals}{A vector containing chemical names in IUPAC notation}

\item{decontaminate}{A logical set to remove uncertain chemicals or junk from output}
}
\value{
A data frame containing chemical names(`chemicals`), their
optimal retention time, exact mass, best identified match factor, and
aggregated component area across every sample it was identified in.
}
\description{
Uses the output from `spreadOut()` and a list of chemicals
to extract data for. Samples that contain a chemical
have all identified area(s) aggregated. The most likely identifications
are determined by prioritizing matches of exact chemical names, followed
by m/z overlaps within precise retention time windows that are determined
by molecular masses published on PubChem.
}
\details{
Communicates with PubChem to collect information on every search
chemical. Uses this information to search the
raw `spreadOut()` data for samples where a chemical exists.
}
\examples{
\dontrun{
query_chemicals = c("Ethyl hexanoate","Methyl salicylate","Octanal","Undecane")
mzExacto(standard_spread, query_chemicals)
}

}
