% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productionHardening.R
\name{plantOccurrenceEvidenceDictionary}
\alias{plantOccurrenceEvidenceDictionary}
\title{Evidence grade dictionary}
\usage{
plantOccurrenceEvidenceDictionary()
}
\value{
Data frame defining canonical plant occurrence evidence grades.
}
\description{
Evidence grade dictionary
}
