% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/published_data.R
\docType{data}
\name{published_data}
\alias{published_data}
\title{Supplemental data set for Tables 1 and 2 in publication}
\format{
A data frame with 2432 rows and 7 columns:
\describe{
 \item{Component.RT}{Retention time of captured component}
 \item{Base.Peak.MZ}{Mass to charge ratio (m/z) of captured component}
 \item{Component.Area}{Area quantified for captured component}
 \item{Compound.Name}{Tentative compound identity for captured component}
 \item{Match.Factor}{Library match factor for captured component (i.e., accuracy of identification)}
 \item{Sample.Name}{Name of sample from which component was captured}
 \item{File.Name}{Name of file from which component was captured}
 }
}
\usage{
published_data
}
\description{
A data frame containing raw output from Agilent's Unknowns Analysis
}
\keyword{datasets}
