% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_spread.R
\docType{data}
\name{standard_spread}
\alias{standard_spread}
\title{Custom Output from `spreadOut()`}
\format{
## `standard_spread`
A list with 7 data frames and 1 list.
\describe{
 \item{Area}{Spread Out Area/Quantity of Chemical}
 \item{Compounds}{Spread Out Name of Tentative Chemicals}
 \item{MZ}{Spread Out Captured M/Z Values}
 \item{MatchFactor}{Spread Out Accuracy of Tentative Matches}
 \item{RT}{Spread Out Retention Times}
 \item{Mass}{Spread Out Compound Mass}
 \item{rtBYmass}{Spread Out ID Codes (Retention Time Pasted to Mass)}
 \item{webInfo}{A list containing all published names, top m/z peaks,
 exact masses, and optimal retention times for every tentative chemical in `standard_data`}
 }
}
\usage{
standard_spread
}
\description{
A list containing all of the output from `spreadOut(standard_data)`.
Has all necessary elements for downstream functions, specifically
`mzExacto()`
}
\keyword{datasets}
