% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorate.R
\name{categorate}
\alias{categorate}
\title{categorate}
\usage{
categorate(compounds, chemical_library, input_format = "wide")
}
\arguments{
\item{compounds}{A vector containing chemical names in IUPAC notation
(preferred).}

\item{chemical_library}{A .CSV file containing columns with chemical
groups. Column names label the group while the rows contain chemicals
that are described by the label.}

\item{input_format}{Designates the data structure for input library
(`chemical_library`). Default setting is "wide" can be changed to "long."}
}
\value{
List with 4 data frames: (1) information from database
searches, (2) atomic and molecular sub-group structure summary, (3) strong
(similarity > 0.95) or moderate (similarity > 0.85) matches with input
library groups, and (4) top chemical from groups a chemical shared a
strong match with.
}
\description{
Searches a vector of chemical names on PubChem for published
information. Uses regular expressions to extract information from 6 data-
bases:
LOTUS the natural products occurrence database [LOTUS](https://lotus.naturalproducts.net/),
Flavor and Extract Manufacturers Association [FEMA](https://www.femaflavor.org/),
Kyoto Encyclopedia of Genes and Genomes [KEGG](https://www.genome.jp/kegg/),
Food and Drug Administration [FDA/SPL](https://www.fda.gov/),
and Reactive Groups from [PubChem](https://pubchem.ncbi.nlm.nih.gov/).
Also downloads structural data in SDF format to summarize the atomic
structures and common molecular groups of each chemical. Finally, the
function downloads structural data for the chemicals within each group
of the input library (`chemical_library`). Each chemical of interest
(`compounds`) is tested against every group of chemicals for structural
overlaps that exceed a threshold of 0.85.
}
\details{
Provides a detailed overview of categorical and structural
information for every chemical of interest. Functional
matches are also generated by matching with groups from a chemical
library.
}
\examples{
compounds = c("3-Octanone","Decane","Mesitylene","1,2,4-trimethyl-benzene",
"D-Limonene","beta-ethyl benzeneethanol","1,4-diethyl benzene",
"1,2-diethyl benzene","1,3,8-p-Menthatriene","(2-methyl-1-propenyl)-Benzene",
"1-Phenyl-1-butene","Linalool","Nonanal","5-nonyl-2-Thiophenecarboxylic acid",
"Dichloroacetaldehyde","Linalyl acetate","Beta-Ocimene")
categorate(compounds, library_data, input_format = "wide")

}
