% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardifyIT.R
\name{standardifyIt}
\alias{standardifyIt}
\title{Post-merge function to quantify compound emission rates relative to internal or external standard(s).}
\usage{
standardifyIt(
  data_in,
  standard_type = "Internal",
  standard_used = "Tetradecane",
  IS_ng = 190.5,
  IS_uL = 1,
  collect_time = 1,
  sample_amt = 1,
  ES_calibration = NA
)
}
\arguments{
\item{input}{uafR 'theMerger()' output.}
}
\description{
If using internal standard (IS), will use the 'standardify()' function with user-specified 
inputs. If using external standard (ES), will require an input "matrix" from which standard curves will 
be derived.
}
