% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standards_model_data.R
\docType{data}
\name{standards_model_data}
\alias{standards_model_data}
\title{Supplemental data set for Figure 1 in publication}
\format{
A data frame with 15 rows and 7 columns:
\describe{
 \item{uafR.RT}{Retention Time identified by uafR for referenced chemical}
 \item{uafR.Mass}{Exact Mass found by uafR for referenced chemical}
 \item{PubChem.Mass}{Mass shown on PubChem web interface for referenced chemical}
 \item{Chemical}{Name of chemical}
 \item{Sample}{Sample from which the chemical was identified}
 \item{ng.Injected}{Amount of standard chemical injected into sample}
 \item{uafR.Peak.Area}{Total component area identified by uafR for referenced chemical}
 }
}
\usage{
standards_model_data
}
\description{
Contains summary data that was used for linear model figures/analyses of
standards data.
}
\keyword{datasets}
